Support writing longs as strings (opt-in)
`jackson-datatype-protobuf` is a missing link in the Jackson ecosystem,
but it doesn't currently provide a way to write 64-bit integers using
their canonical representation, which poses a problem for strict
downstream parsers given the limitation of the maximum value to
[2^53 - 1](https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Number/MAX_SAFE_INTEGER)
([alt](https://developers.google.com/discovery/v1/type-format)) in
JavaScript, whereas this value is for instance
[2^63 - 1](https://docs.oracle.com/javase/tutorial/java/nutsandbolts/datatypes.html)
in Java.

The change proposed here is to provide users who need it with an option
to activate this canonical representation, without altering the existing
default behavior.